requirements.txt
Flask>=2.0.0
Jinja2==3.1.3
Werkzeug==3.0.1
itsdangerous==2.1.2
MarkupSafe==2.1.5
requests==2.32.3
pandas==2.2.2
python-dotenv>=1.0.0
urllib3==1.26.13
python-dateutil==2.9.0.post0
gunicorn
Authlib>=1.0.0
Flask-Login>=0.6.2
